Move PropertyGroup to Rf.hpp
